+++ Changelog.md
@@ -6,7 +6,7 @@
 - Add workaround for CSP rules in Safari (#5526)
 - Display smaller error message when nbconvert fails to avoid a cookie overflow (#5510)
 - Change level mark input field to accept change increments of 0.01 (#5546)
-
+- Add test results download modal for assignment summary (#5561)
 
 ## [v1.13.2]
 - Ensure "Create all groups" button uses existing repos if they already exist (#5504)